Use  `setupFilesAfterEnv`
